requirements.txt
requests==2.23.0
pillow